refactor: move src/logger.js -> src/reporters/logger.js
