+++ src/reporters/apiary-reporter.js
@@ -3,7 +3,7 @@ const generateUuid = require('uuid/v4');
 const os = require('os');
 const request = require('request');
 
-const logger = require('./../logger');
+const logger = require('./logger');
 const packageData = require('./../../package.json');
 
 const CONNECTION_ERRORS = [
