+++ src/reporters/cli-reporter.js
@@ -1,4 +1,4 @@
-const logger = require('./../logger');
+const logger = require('./logger');
 const prettifyResponse = require('./../prettify-response');
 
 function CliReporter(emitter, stats, tests, inlineErrors, details) {
