+++ src/reporters/dot-reporter.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('./logger');
 const prettifyResponse = require('../prettify-response');
 
 function DotReporter(emitter, stats, tests) {
