+++ src/reporters/markdown-reporter.js
@@ -6,7 +6,7 @@ const file = require('file');
 const fsExtra = require('fs-extra');
 const pathmodule = require('path');
 
-const logger = require('./../logger');
+const logger = require('./logger');
 const prettifyResponse = require('./../prettify-response');
 
 function MarkdownReporter(emitter, stats, tests, path, details) {
