+++ test/integration/cli/configuration-cli-test.js
@@ -4,7 +4,7 @@ const proxyquire = require('proxyquire').noCallThru();
 const sinon = require('sinon');
 const { assert } = require('chai');
 
-const loggerStub = require('../../../src/logger');
+const loggerStub = require('../../../src/reporters/logger');
 const configUtils = require('../../../src/config-utils');
 
 const PORT = 9876;
@@ -15,17 +15,17 @@ let stderr = '';
 let stdout = '';
 
 const addHooksStub = proxyquire('../../../src/add-hooks', {
-  './logger': loggerStub
+  './reporters/logger': loggerStub
 });
 
 const transactionRunner = proxyquire('../../../src/transaction-runner', {
   './add-hooks': addHooksStub,
-  './logger': loggerStub
+  './reporters/logger': loggerStub
 });
 
 const dreddStub = proxyquire('../../../src/dredd', {
   './transaction-runner': transactionRunner,
-  './logger': loggerStub
+  './reporters/logger': loggerStub
 });
 
 const CLIStub = proxyquire('../../../src/cli', {
